refactor: move function for update USER cache after avatar mytation to separate file
