+++ src/graphql/mutations/avatar/avatar.cache.ts
@@ -1,19 +1,7 @@
-import { ApolloCache, gql, NormalizedCacheObject } from '@apollo/client';
-import { IAvatarReturn } from '../../interfaces/IAvatar';
-import { IUserAllResult } from '../../interfaces/IUser.interface';
-import { USER } from '../queries/user';
-
-export const UPLOAD_AVATAR = gql`
-  mutation UploadAvatar($id: ID!, $avatar: AvatarInput!) {
-    uploadAvatar(id: $id, avatar: $avatar)
-  }
-`;
-
-export const DELETE_AVATAR = gql`
-  mutation DeleteAvatar($id: ID!) {
-    deleteAvatar(id: $id)
-  }
-`;
+import { ApolloCache, NormalizedCacheObject } from '@apollo/client';
+import { IAvatarReturn } from '../../../interfaces/IAvatar';
+import { IUserAllResult } from '../../../interfaces/IUser.interface';
+import { USER } from '../../queries/user';
 
 export const updateUserCacheAfterAvatarMutation = (
   cache: ApolloCache<NormalizedCacheObject>,
